+++ raster_feeder/config.py
@@ -21,6 +21,11 @@ LOG_DIR = join(BUILDOUT_DIR, 'var', 'log')
 REDIS_HOST = 'localhost'
 REDIS_DB = 0
 
+# Lizard API credentials
+LIZARD_USERNAME = 'override'
+LIZARD_PASSWORD = 'override'
+LIZARD_TEMPLATE = 'override'
+
 # Import local settings
 try:
     from .localconfig import *  # NOQA